+++ CobwebApp/src/main/java/cobweb3d/plugins/exchange/utility/Calculation.java
@@ -1,5 +1,5 @@
 package cobweb3d.plugins.exchange.utility;
 
 public interface Calculation {
-    int calculateU(int x, int y, float A, float B);
+    float calculateU(float x, float y, float A, float B);
 }
