+++ CobwebApp/src/main/java/cobweb3d/plugins/exchange/utility/UtilityFunctionParam.java
@@ -32,8 +32,8 @@ public class UtilityFunctionParam implements ParameterSerializable {
         NONE("0", "None", ((x, y, A, B) -> 0)),
         SQRTxy("1", "\u221A(x * y)", ((x, y, A, B) -> x * y)),
         MINxy("2", "min(x, y)", ((x, y, A, B) -> x < y ? x : y)),
-        xPaMyPb("3", "x^A * y^B", ((x, y, A, B) -> (int) (Math.pow(x, A) * Math.pow(y, B)))),
-        axAby("4", "Ax + By", ((x, y, A, B) -> (int) ((A * x) + (B * y))));
+        xPaMyPb("3", "x^A * y^B", ((x, y, A, B) -> (float) (Math.pow(x, A) * Math.pow(y, B)))),
+        axAby("4", "Ax + By", ((x, y, A, B) -> ((A * x) + (B * y))));
 
         private final String id;
         private final String name;
@@ -72,7 +72,7 @@ public class UtilityFunctionParam implements ParameterSerializable {
         }
 
         @Override
-        public int calculateU(int x, int y, float A, float B) {
+        public float calculateU(float x, float y, float A, float B) {
             return calculation.calculateU(x, y, A, B);
         }
     }
